python-backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
rembg>=2.0.60
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
onnxruntime>=1.15.0
beautifulsoup4>=4.12.0
